requirements.txt
supabase>=2.25.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
